fix(examples): Increase scale-to-zero cooldown for Wordpress
The current scale-to-zero cooldown (`1000ms`) in insufficient to fully
start the Wordpress instance. With this, the first queries fail, as the
service is not fully started. Simply put, the instance goes to sleep to
soon after being started and, when it is woken up by a query, it cannot
serve it.

Update the scale-to-zero cooldown to `3000ms`.

Signed-off-by: Razvan Deaconescu <razvand@unikraft.io>