fix: Copy the proper dynamic linker through the symlink target
The interpreter/linker's path defined by the tested binary is
`/lib64/ld-linux-x86-64.so.2`. In the container this is actually
a symlink to `/lib/x86_64-linux-gnu/ld-linux-x86-64.so.2`, but
the target would have been the one copied anyway.

Regardless, the interpreter we pick in Unikraft's elfloader is the
one we extract from the ELF headers which has the former path.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>