feat(postgres): Add patch for running as root
With the release of the new metros, supporting virtiofs, we currently do
not have a way to support filesystem ownership.

This commit builds a patched version of Postgres that allows execution of root
user inside the VM.

It also improves the initrd size, as a consequence of building the filesystem
from scratch.

Signed-off-by: Alex-Andrei Cioc <andrei.cioc@unikraft.io>